requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
python-multipart==0.0.12
pillow==10.4.0
opencv-python==4.10.0.84
numpy==2.0.2
torch==2.5.1
torchvision==0.20.1
